+++ src/components/nav/components/desktop/components/action_bar/components/setting_list/index.tsx
@@ -17,7 +17,7 @@ import {
 } from '@material-ui/core';
 import { Close as CloseIcon } from '@material-ui/icons';
 import {
-  THEME_LIST, DATE_LIST, TX_LIST,
+  THEME_LIST, DATE_LIST,
 } from '@recoil/settings';
 import { useSettingList } from './hooks';
 import { useStyles } from './styles';
@@ -146,7 +146,7 @@ const Settings: React.FC<{
               </Select>
             </div>
 
-            <div className={classes.formItem}>
+            {/* <div className={classes.formItem}>
               <Typography className="form-item--label">
                 {t('txListFormat')}
               </Typography>
@@ -168,7 +168,7 @@ const Settings: React.FC<{
                     </MenuItem>
                   ))}
               </Select>
-            </div>
+            </div> */}
           </form>
         </DialogContent>
         <DialogActions>